Start adding widgets to Booking
Started adding widgets to the dialog, no functionality yet,
just laying things out.

In this step I just added a calendar and 2 QDateEdit
with a button and a label each. This button will later
be used to show a pop up calendar to pick a date,
and the calendar that's always visible will be
used to show the picked dates iterval.