Add new page to CheckIn
Added the page VariantData to the ChekIn dialog, it is meant to take
the data that keeps changing with every visit, even if the guest
has it's data already registered.
When the guest is already registered this is the page shown by default
and all its data is pre-filled automatically.